shader_recompiler: TXQ: Skip QueryLevels when possible by ammerj
"TXQ has a mask for the components that are to be written out. This PR skips emitting the shader code to query the number of mip levels if that component is disabled in the mask.
This is done in preparation for fixing TXQ with MSAA textures. MSAA textures do not have mip maps, so shader compilations will fail if there's an attempt to query the number of mip levels of an MSAA texture.
"

https://github.com/yuzu-emu/yuzu/pull/9694