shaders: Fix TXQ with MSAA textures by ameerj
"TXQ can operate on MSAA textures, but has no flag within the instruction indicating if that is the case. The shader backends require knowing if the texture is MSAA as they cannot query LOD related information for MSAA textures.

This implementation relies on an assumption that TLD/texelFetch is called on the MSAA texture elsewhere in the shader, which can track if the texture binding is MSAA.
This assumption will fall apart if there are shaders which query the size of the MSAA texture, but never sample from it."

https://github.com/yuzu-emu/yuzu/pull/9703